Added state refreshes for when user used CRUD routes.

+++ server/sql/subscriptions/getSubscriptionsByUserId.sql
@@ -3,7 +3,7 @@ SELECT
   nickname, 
   to_char(due_date, 'YYYY-MM-DD') AS due_date, 
   reminder_days, 
-  amount, 
+  amount::numeric, 
   subscription_uuid 
 FROM subscriptions 
 WHERE(user_id = ${userId});
